Revert "MULE-9855 Make MuleMessage immutable. Step 2"
