+++ core/src/main/java/org/mule/runtime/core/message/DefaultMuleMessageBuilderFactory.java
@@ -7,6 +7,7 @@
 package org.mule.runtime.core.message;
 
 import org.mule.runtime.api.message.AbstractMuleMessageBuilderFactory;
+import org.mule.runtime.core.DefaultMuleMessageBuilder;
 import org.mule.runtime.core.api.MuleMessage;
 
 import java.io.Serializable;
